Add ability to change foreground font based on contrast with background.
Sometimes foreground colors and the color being used for the focus / highlight
attributes are too close, making it hard to read.  As a workaround,
get_palette_entries now takes a list of entires that can have their foreground
colors changed to increase contrast.  This is only done for high color mode,
since there's ample contrast between all of the 16 baic colors, and is currently
only being used to turn dark shades to black due to the hard-coded focus and
highlight colors.  At some point, if theming / customization of colors is
implemented, we'll want to make this more flexible.