Add always_visible toggle to scrollbar
When the scrollbar becomes visible, the canvas of the original widget
may need to be re-rendered in order to cope with having e.g. one less
column.

For items like urwid.Text, this re-flow means the text will jump around
once the ScrollBar becomes visible. A simple adjustment here is to allow
the ScrollBar to always be visible, so that the text does not jump when
it crosses the "scrollable length" threshold.

Add a new initializer keyword that defaults to the old behavior.